Bug 1198613 - Plugin management: Framework for a new screen, splashscreen or editor is not saved

+++ uberfire-extensions/uberfire-runtime-plugins/uberfire-runtime-plugins-backend/src/main/java/org/uberfire/ext/plugin/backend/PluginServicesImpl.java
@@ -316,8 +316,10 @@ public class PluginServicesImpl implements PluginServices {
     }
 
     private void clearDirectory(Path directory) {
-        for (Path path : ioService.newDirectoryStream(directory)) {
-            boolean b = ioService.deleteIfExists(path);
+        if (ioService.exists(directory)) {
+            for (Path path : ioService.newDirectoryStream(directory)) {
+                boolean b = ioService.deleteIfExists(path);
+            }
         }
     }
